requirements/common.txt
# Django
Django==2.0.1
django-extensions==1.9.9
itypes==1.1.0
packaging==16.8
django-cors-headers==1.2.0
pyparsing==2.2.0
requests==2.18.4
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
responses==0.8.1
whitenoise==4.0b4

 # DRF
 djangorestframework==3.7.7

 #Django filters
django-filter==1.1.0

 # DB connector
psycopg2==2.7.3.2

# Swagger
coreapi==2.3.3
openapi-codec==1.3.2
django-rest-swagger

# Testing
coverage==4.4.2
pytest-django
